Added test for vertical wins, fixed bug in pivot.
Creating the pivot table with the list multi-
plication method actually resulted in one
list with 3 references to it, rather than 3
distinct lists. When marking a cell in
one list (row), the same cell would be
marked in all lists.

Corrected this by making 3 separate lists
before doing the pivot.